Move `fix_absolute_paths` to main package
