+++ post/clang_tidy_review/clang_tidy_review/__init__.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: MIT
 # See LICENSE for more information
 
-import argparse
 import fnmatch
 import itertools
 import json
@@ -481,6 +480,33 @@ def try_relative(path):
         return pathlib.Path(path).relative_to(here)
     except ValueError:
         return pathlib.Path(path).resolve()
+
+
+def fix_absolute_paths(build_compile_commands, base_dir):
+    """Update absolute paths in compile_commands.json to new location, if
+    compile_commands.json was created outside the Actions container
+    """
+
+    basedir = pathlib.Path(base_dir).resolve()
+    newbasedir = pathlib.Path(".").resolve()
+
+    if basedir == newbasedir:
+        return
+
+    print(f"Found '{build_compile_commands}', updating absolute paths")
+    # We might need to change some absolute paths if we're inside
+    # a docker container
+    with open(build_compile_commands, "r") as f:
+        compile_commands = json.load(f)
+
+    print(f"Replacing '{basedir}' with '{newbasedir}'", flush=True)
+
+    modified_compile_commands = json.dumps(compile_commands).replace(
+        str(basedir), str(newbasedir)
+    )
+
+    with open(build_compile_commands, "w") as f:
+        f.write(modified_compile_commands)
 
 
 def format_notes(notes, offset_lookup):
